6_outbox_with_nats/pyproject.toml
[tool.poetry]
name = "outbox"
version = "0.1.0"
description = ""
authors = ["ksimmi <ksimmi@yandex.com>"]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.2"
nats-py = "^2.1.0"
sqlalchemy = {git = "https://github.com/sqlalchemy/sqlalchemy.git", rev = "main"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
